+++ src/index.html
@@ -18,19 +18,19 @@
 
     <h1>Link Up!</h1>
     
-
+  <div id = "register_inputs" class="register_entry"></div>
 
     <div class="down">
-    <a href = "#btn1">REGISTRARSE</a>
-   <a href = "#btn2">INICIAR SESIÓN</a>
+<button id = "register">REGISTRARSE</button>
+   <a href = "login">INICIAR SESIÓN</a>
     </div>
     
     
 
       
 
-
-  <script src="data.js"></script>
+<script src="views/pages/Register.js"></script>
+  <script src="app.js"></script>
   <script src="main.js"></script>
 
 </body>
