+++ src/register.html
@@ -1,0 +1,28 @@
+<!--<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Document</title>
+</head>
+<body>
+    <p>
+    <label class="lab">Email</label>
+    <br>
+   <input class= "input" id="email" type ="email" ></p>
+   <p>
+   <label class="lab" >Password</label>
+   <br>
+   <input class= "input" id="pass" type ="password" ></p>
+  
+
+   <button id="register" onclick="register()">Enviar</button>
+   <button id="register">Registrarse</button>
+   <section id="user-data">
+     
+<button id="logout">Cerrar Sesión</button>
+
+   </section>
+</body>
+</html>-->
