+++ src/style.css
@@ -18,7 +18,7 @@
     
 }
 
-a{
+#register{
 
     font-family: 'Roboto Condensed', sans-serif;
     color: white;
