+++ src/views/home.js
@@ -1,0 +1,11 @@
+import buttoRegister from '../components/buttoRegister.js'
+
+const Home = (father,callback) => {
+buttonRegister('registrarse', father).render ();
+const container = document. createElement (section);
+father.appendChild(container);
+buttonRegister('registrarse', callback).render(father);{
+
+};
+export default Home
+}
